File reconstruction & db integration handling Moved folders: - routeHandlers/ - routers/ - methods/ into [private](./private) in order to properly use it, as it was under utilized folder.
Created a new folder inside private:
- db/
As it is to handle database specific javascript integration. In the future: could/should move all db queries into this folder for seperating logic properly, maybe a class?
Added a specific file: connection.js, to handle initial db connection.

General reporpuse change:
- loginHander.js
- signupHandler.js
- userHandler.js
Reverted these back into handling (req, res) parameters alongside handling them as async. Upon necessity of using DB functions, requires the import of db instead of having more parameters (can become very tedious real fast).
