+++ private/db/connection.js
@@ -1,0 +1,19 @@
+import mysql from 'mysql2';
+import 'dotenv/config';
+
+const db = mysql.createConnection({
+    host: process.env.DB_HOST,
+    user: process.env.DB_USER,
+    password: process.env.DB_PASSWORD,
+    database: process.env.DB_NAME,
+});
+
+db.connect((err) => {
+    if (err) {
+        console.error('Database connection error:', err);
+    } else {
+        console.log('Connected to MySQL');
+    }
+});
+
+export default db;
