+++ private/methods/responseMethods.js
