+++ private/methods/utilsMethods.js
