+++ private/routeHandlers/auth/rewriteForLogin.js
